Use INTEGER instead of MEDIUMINT (#3309)

+++ connectors/src/lib/models/confluence.ts
@@ -147,7 +147,7 @@ ConfluencePage.init(
       defaultValue: DataTypes.NOW,
     },
     version: {
-      type: DataTypes.MEDIUMINT,
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
     skipReason: {
